politigraph-ocr-votes-log/politigraph-votes-log-extractor/pyproject.toml
[project]
name = "politigraph_votes_extractor"
version = "0.1.0"
dependencies = [
    "requests", "pandas", "numpy", 
    "opencv_python", "pdf2image", "Pillow",
    "rapidfuzz==3.13.0",
    "easyocr"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
package = true